+++ mg_custom_nodes/KLL535_ComfyUI_Simple_Qwen3-VL-gguf_20260516/__init__.py
@@ -2,7 +2,7 @@
 #__init__.py
 
 from .qwen3vl_node import SimpleQwen3VL_GGUF_Node
-from .utils_node import MasterPromptLoader,SimpleStyleSelector,SimpleCameraSelector,UnloadQwenModel,SimpleRemoveThinkNode,SimpleTriggerNode
+from .utils_node import MasterPromptLoader,SimpleStyleSelector,SimpleCameraSelector,UnloadQwenModel,SimpleRemoveThinkNode,SimpleTriggerNode,TextToBatchNode,SimpleTextInsertNode,SimpleTextReplaceNode,SimpleJoinStringsNode
 from .deprecated_node import Qwen3VL_GGUF_Node
 from .configurator import Qwen3VL_ModelConfig, Qwen3VL_SamplingConfig
 
@@ -16,7 +16,10 @@ NODE_CLASS_MAPPINGS = {
     "SimpleTriggerNode": SimpleTriggerNode,
     "Qwen3VL_ModelConfig": Qwen3VL_ModelConfig,
     "Qwen3VL_SamplingConfig": Qwen3VL_SamplingConfig,
-
+    "SimpleTextToBatchNode": TextToBatchNode,
+    "SimpleTextInsertNode": SimpleTextInsertNode,
+    "SimpleTextReplaceNode": SimpleTextReplaceNode,
+    "SimpleJoinStringsNode": SimpleJoinStringsNode,
 
     #deprecated_node
     "SimpleQwenVLgguf": Qwen3VL_GGUF_Node,
@@ -32,6 +35,10 @@ NODE_DISPLAY_NAME_MAPPINGS = {
     "SimpleTriggerNode": "Simple Trigger Node",
     "Qwen3VL_ModelConfig": "LLM Model Config",
     "Qwen3VL_SamplingConfig": "LLM Sampling Config",
+    "SimpleTextToBatchNode": "Simple Text To Batch",
+    "SimpleTextInsertNode": "Simple Text Insert",
+    "SimpleTextReplaceNode": "Simple Text Replace",
+    "SimpleJoinStringsNode": "Simple Join Strings",
 
     #deprecated_node
     "SimpleQwenVLgguf": "Qwen-VL Vision Language Model",